[backend] Move room guards and decorators to their own folders
